pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pykmstool"
version = "2.0.0"
description = "Google Cloud KMS Certificate Signing Request (CSR) Generation Tool"
readme = "README.md"
authors = [
    {name = "Michał Leszczyński", email = "ml@icedev.pl"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "cryptography>=46.0.4",
    "google-cloud-resource-manager>=1.16.0",
    "google-cloud-kms>=3.10.0",
    "crcmod>=1.7",
    "click>=8.3.1"
]

[project.scripts]
pykmstool = "pykmstool.cli:cli"

[project.urls]
Homepage = "https://github.com/icedevml/pykmstool"
Repository = "https://github.com/icedevml/pykmstool"
Issues = "https://github.com/icedevml/pykmstool/issues"

[tool.setuptools.packages.find]
where = ["src"]
